create libcontainer handler for common code
