+++ container/crio/handler.go
@@ -21,7 +21,6 @@ import (
 	"path/filepath"
 	"strconv"
 	"strings"
-	"time"
 
 	"github.com/google/cadvisor/container"
 	"github.com/google/cadvisor/container/common"
@@ -29,47 +28,31 @@ import (
 	"github.com/google/cadvisor/fs"
 	info "github.com/google/cadvisor/info/v1"
 
-	"github.com/opencontainers/runc/libcontainer/cgroups"
 	cgroupfs "github.com/opencontainers/runc/libcontainer/cgroups/fs"
 	libcontainerconfigs "github.com/opencontainers/runc/libcontainer/configs"
 )
 
 type crioContainerHandler struct {
-	name               string
-	id                 string
-	aliases            []string
 	machineInfoFactory info.MachineInfoFactory
 
 	// Absolute path to the cgroup hierarchies of this container.
 	// (e.g.: "cpu" -> "/sys/fs/cgroup/cpu/test")
 	cgroupPaths map[string]string
 
-	// Manager of this container's cgroups.
-	cgroupManager cgroups.Manager
-
 	// the CRI-O storage driver
 	storageDriver    storageDriver
 	fsInfo           fs.FsInfo
 	rootfsStorageDir string
 
-	// Time at which this container was created.
-	creationTime time.Time
-
 	// Metadata associated with the container.
-	labels map[string]string
 	envs   map[string]string
+	labels map[string]string
 
 	// TODO
 	// crio version handling...
 
-	// The container PID used to switch namespaces as required
-	pid int
-
 	// Image name used for this container.
 	image string
-
-	// The host root FS to read
-	rootFs string
 
 	// The network mode of the container
 	// TODO
@@ -84,6 +67,10 @@ type crioContainerHandler struct {
 
 	// container restart count
 	restartCount int
+
+	reference info.ContainerReference
+
+	libcontainerHandler *containerlibcontainer.Handler
 }
 
 var _ container.ContainerHandler = &crioContainerHandler{}
@@ -150,25 +137,29 @@ func newCrioContainerHandler(
 		rootfsStorageDir = filepath.Join(rootfsStorageDir, "diff")
 	}
 
-	// TODO: extract object mother method
-	handler := &crioContainerHandler{
-		id:                 id,
-		name:               name,
-		machineInfoFactory: machineInfoFactory,
-		cgroupPaths:        cgroupPaths,
-		cgroupManager:      cgroupManager,
-		storageDriver:      storageDriver,
-		fsInfo:             fsInfo,
-		rootFs:             rootFs,
-		rootfsStorageDir:   rootfsStorageDir,
-		envs:               make(map[string]string),
-		ignoreMetrics:      ignoreMetrics,
+	containerReference := info.ContainerReference{
+		Id:        id,
+		Name:      name,
+		Aliases:   []string{cInfo.Name, id},
+		Namespace: CrioNamespace,
 	}
 
-	handler.creationTime = time.Unix(0, cInfo.CreatedTime)
-	handler.pid = cInfo.Pid
-	handler.aliases = append(handler.aliases, cInfo.Name, id)
-	handler.labels = cInfo.Labels
+	libcontainerHandler := containerlibcontainer.NewHandler(cgroupManager, rootFs, cInfo.Pid, ignoreMetrics)
+
+	// TODO: extract object mother method
+	handler := &crioContainerHandler{
+		machineInfoFactory:  machineInfoFactory,
+		cgroupPaths:         cgroupPaths,
+		storageDriver:       storageDriver,
+		fsInfo:              fsInfo,
+		rootfsStorageDir:    rootfsStorageDir,
+		envs:                make(map[string]string),
+		labels:              cInfo.Labels,
+		ignoreMetrics:       ignoreMetrics,
+		reference:           containerReference,
+		libcontainerHandler: libcontainerHandler,
+	}
+
 	handler.image = cInfo.Image
 	// TODO: we wantd to know graph driver DeviceId (dont think this is needed now)
 
@@ -204,12 +195,7 @@ func (self *crioContainerHandler) Cleanup() {
 }
 
 func (self *crioContainerHandler) ContainerReference() (info.ContainerReference, error) {
-	return info.ContainerReference{
-		Id:        self.id,
-		Name:      self.name,
-		Aliases:   self.aliases,
-		Namespace: CrioNamespace,
-	}, nil
+	return self.reference, nil
 }
 
 func (self *crioContainerHandler) needNet() bool {
@@ -285,7 +271,7 @@ func (self *crioContainerHandler) getFsStats(stats *info.ContainerStats) error {
 }
 
 func (self *crioContainerHandler) GetStats() (*info.ContainerStats, error) {
-	stats, err := containerlibcontainer.GetStats(self.cgroupManager, self.rootFs, self.pid, self.ignoreMetrics)
+	stats, err := self.libcontainerHandler.GetStats()
 	if err != nil {
 		return stats, err
 	}
@@ -314,7 +300,7 @@ func (self *crioContainerHandler) ListContainers(listType container.ListType) ([
 func (self *crioContainerHandler) GetCgroupPath(resource string) (string, error) {
 	path, ok := self.cgroupPaths[resource]
 	if !ok {
-		return "", fmt.Errorf("could not find path for resource %q for container %q\n", resource, self.name)
+		return "", fmt.Errorf("could not find path for resource %q for container %q\n", resource, self.reference.Name)
 	}
 	return path, nil
 }
@@ -328,7 +314,7 @@ func (self *crioContainerHandler) GetContainerIPAddress() string {
 }
 
 func (self *crioContainerHandler) ListProcesses(listType container.ListType) ([]int, error) {
-	return containerlibcontainer.GetProcesses(self.cgroupManager)
+	return self.libcontainerHandler.GetProcesses()
 }
 
 func (self *crioContainerHandler) Exists() bool {
