+++ container/rkt/handler.go
@@ -28,41 +28,22 @@ import (
 	"golang.org/x/net/context"
 
 	"github.com/golang/glog"
-	"github.com/opencontainers/runc/libcontainer/cgroups"
 	cgroupfs "github.com/opencontainers/runc/libcontainer/cgroups/fs"
 	"github.com/opencontainers/runc/libcontainer/configs"
 )
 
 type rktContainerHandler struct {
-	rktClient rktapi.PublicAPIClient
-	// Name of the container for this handler.
-	name               string
-	cgroupSubsystems   *libcontainer.CgroupSubsystems
 	machineInfoFactory info.MachineInfoFactory
 
 	// Absolute path to the cgroup hierarchies of this container.
 	// (e.g.: "cpu" -> "/sys/fs/cgroup/cpu/test")
 	cgroupPaths map[string]string
 
-	// Manager of this container's cgroups.
-	cgroupManager cgroups.Manager
-
-	// Whether this container has network isolation enabled.
-	hasNetwork bool
-
 	fsInfo fs.FsInfo
-
-	rootFs string
 
 	isPod bool
 
-	aliases []string
-
-	pid int
-
 	rootfsStorageDir string
-
-	labels map[string]string
 
 	// Filesystem handler.
 	fsHandler common.FsHandler
@@ -70,6 +51,12 @@ type rktContainerHandler struct {
 	ignoreMetrics container.MetricSet
 
 	apiPod *rktapi.Pod
+
+	labels map[string]string
+
+	reference info.ContainerReference
+
+	libcontainerHandler *libcontainer.Handler
 }
 
 func newRktContainerHandler(name string, rktClient rktapi.PublicAPIClient, rktPath string, cgroupSubsystems *libcontainer.CgroupSubsystems, machineInfoFactory info.MachineInfoFactory, fsInfo fs.FsInfo, rootFs string, ignoreMetrics container.MetricSet) (container.ContainerHandler, error) {
@@ -122,30 +109,27 @@ func newRktContainerHandler(name string, rktClient rktapi.PublicAPIClient, rktPa
 		Paths: cgroupPaths,
 	}
 
-	hasNetwork := false
-	if isPod {
-		hasNetwork = true
-	}
+	libcontainerHandler := libcontainer.NewHandler(cgroupManager, rootFs, pid, ignoreMetrics)
 
 	rootfsStorageDir := getRootFs(rktPath, parsed)
 
+	containerReference := info.ContainerReference{
+		Name:      name,
+		Aliases:   aliases,
+		Namespace: RktNamespace,
+	}
+
 	handler := &rktContainerHandler{
-		name:               name,
-		rktClient:          rktClient,
-		cgroupSubsystems:   cgroupSubsystems,
-		machineInfoFactory: machineInfoFactory,
-		cgroupPaths:        cgroupPaths,
-		cgroupManager:      cgroupManager,
-		fsInfo:             fsInfo,
-		hasNetwork:         hasNetwork,
-		rootFs:             rootFs,
-		isPod:              isPod,
-		aliases:            aliases,
-		pid:                pid,
-		labels:             labels,
-		rootfsStorageDir:   rootfsStorageDir,
-		ignoreMetrics:      ignoreMetrics,
-		apiPod:             apiPod,
+		machineInfoFactory:  machineInfoFactory,
+		cgroupPaths:         cgroupPaths,
+		fsInfo:              fsInfo,
+		isPod:               isPod,
+		rootfsStorageDir:    rootfsStorageDir,
+		ignoreMetrics:       ignoreMetrics,
+		apiPod:              apiPod,
+		labels:              labels,
+		reference:           containerReference,
+		libcontainerHandler: libcontainerHandler,
 	}
 
 	if !ignoreMetrics.Has(container.DiskUsageMetrics) {
@@ -174,11 +158,7 @@ func createLabels(annotations []*rktapi.KeyValue) map[string]string {
 }
 
 func (handler *rktContainerHandler) ContainerReference() (info.ContainerReference, error) {
-	return info.ContainerReference{
-		Name:      handler.name,
-		Aliases:   handler.aliases,
-		Namespace: RktNamespace,
-	}, nil
+	return handler.reference, nil
 }
 
 func (handler *rktContainerHandler) Start() {
@@ -190,7 +170,7 @@ func (handler *rktContainerHandler) Cleanup() {
 }
 
 func (handler *rktContainerHandler) GetSpec() (info.ContainerSpec, error) {
-	hasNetwork := handler.hasNetwork && !handler.ignoreMetrics.Has(container.NetworkUsageMetrics)
+	hasNetwork := handler.isPod && !handler.ignoreMetrics.Has(container.NetworkUsageMetrics)
 	hasFilesystem := !handler.ignoreMetrics.Has(container.DiskUsageMetrics)
 
 	spec, err := common.GetSpec(handler.cgroupPaths, handler.machineInfoFactory, hasNetwork, hasFilesystem)
@@ -242,7 +222,7 @@ func (handler *rktContainerHandler) getFsStats(stats *info.ContainerStats) error
 }
 
 func (handler *rktContainerHandler) GetStats() (*info.ContainerStats, error) {
-	stats, err := libcontainer.GetStats(handler.cgroupManager, handler.rootFs, handler.pid, handler.ignoreMetrics)
+	stats, err := handler.libcontainerHandler.GetStats()
 	if err != nil {
 		return stats, err
 	}
@@ -274,7 +254,7 @@ func (self *rktContainerHandler) GetContainerIPAddress() string {
 func (handler *rktContainerHandler) GetCgroupPath(resource string) (string, error) {
 	path, ok := handler.cgroupPaths[resource]
 	if !ok {
-		return "", fmt.Errorf("could not find path for resource %q for container %q\n", resource, handler.name)
+		return "", fmt.Errorf("could not find path for resource %q for container %q\n", resource, handler.reference.Name)
 	}
 	return path, nil
 }
@@ -284,11 +264,11 @@ func (handler *rktContainerHandler) GetContainerLabels() map[string]string {
 }
 
 func (handler *rktContainerHandler) ListContainers(listType container.ListType) ([]info.ContainerReference, error) {
-	return common.ListContainers(handler.name, handler.cgroupPaths, listType)
+	return common.ListContainers(handler.reference.Name, handler.cgroupPaths, listType)
 }
 
 func (handler *rktContainerHandler) ListProcesses(listType container.ListType) ([]int, error) {
-	return libcontainer.GetProcesses(handler.cgroupManager)
+	return handler.libcontainerHandler.GetProcesses()
 }
 
 func (handler *rktContainerHandler) Exists() bool {
